Fix warning: cast from pointer to integer of different size.

+++ tools/mbspc/botlib/l_struct.h
@@ -29,6 +29,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  *
  *****************************************************************************/
 
+#include <stddef.h>
 
 #define MAX_STRINGFIELD				80
 //field types
@@ -48,7 +49,7 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 typedef struct fielddef_s
 {
 	char *name;										//name of the field
-	int offset;										//offset in the structure
+	size_t offset;										//offset in the structure
 	int type;										//type of the field
 	//type specific fields
 	int maxarray;									//maximum array size
